Revised the way the sound factory knows whether a sound driver exists - also doubled XA2 queued buffers
The first change is made to decouple the SoundFactory implementation
from the specific drivers.  Simply excluding a driver from the build
will now no longer make it available for selection.  The second change
is one I hope fixing some issues for @Frank-74 .  By doubling the number
of buffers available in XA2 implementation we shouldn't run into an
issue with the buffers depleting completely.  Wait and see...